backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
SQLAlchemy==2.0.35
PyMySQL==1.1.1
cryptography==43.0.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Image Processing & OCR
Pillow==10.4.0
easyocr==1.7.2
opencv-python==4.10.0.84
numpy==1.26.4
torch==2.5.1
torchvision==0.20.1

# QR Code Generation
qrcode[pil]==8.0

# Utilities
jinja2==3.1.4
python-dateutil==2.9.0

# Database Migration
alembic==1.13.3
